AccountId mod and tests
